requirements.txt
crewai==0.28.8
streamlit==1.32.0
google-generativeai
python-dotenv==1.0.1
